+++ NEWS.md
@@ -7,6 +7,7 @@
 - fix ROI writing
 - `EasyFITS` replaced by `AstroFITS`.
 - `MultivariateOnlineStatistics` replaced by `OnlineSampleStatistics`.
+- Add I/O methods for `CalibrationData` to/from FITS files
 
 ## Version v0.3.5